+++ mission03/feature/valid/condition.js
@@ -18,3 +18,15 @@ export const isMatchPassword = (input) => {
   const passwordForm = form.querySelector("#password"); // form 안의 #password 필드 찾기
   return input.value.trim() === passwordForm.value.trim();
 };
+
+export const isAllValid = (inputs) => {
+  // 모든 input 요소가 유효한지 체크하는 함수
+  for (const input of inputs) {
+    const field = input.closest(".field"); // 가까운 field 요소 가져오기
+    const errorMessage = field.querySelector(".error-message"); // field의 에러 메시지 가져오기
+
+    if (isEmpty(input) || errorMessage) return false; // 비어있거나 오류 메시지가 있으면 유효하지 않음
+  }
+
+  return true; // 모든 요소에 값이 있고 오류 메시지가 없으면 유효함
+};
